Move the Annotation and Document models to the h.api package
These are the concern of the API package. This commit moves them there,
but continues to expose them in `h.models` as well for convenience's
sake.
